+++ Cargo.toml
@@ -62,6 +62,7 @@ tendermint = "0.32"
 tendermint-config = "0.32"
 tendermint-proto = "0.32"
 tendermint-rpc = "0.32"
+thiserror = "1"
 tokio = "1.28"
 tonic = "0.9"
 tracing = "0.1"
